% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inst.all.R
\name{inst.all}
\alias{inst.all}
\title{A inst.all Function}
\usage{
inst.all(pckgName, pathGit = NULL, bioc = FALSE, proxy = FALSE,
  urlproxy = "conf_www.ir.vhebron.net", portproxy = 8081,
  force.install = FALSE, ...)
}
\arguments{
\item{pckgName}{names of package}

\item{pathGit}{If Github package: Repository address in the format username/repo[/subdir] .}

\item{bioc}{TRUE if bioconductor package. Default value is FALSE.}

\item{proxy}{If Github package: TRUE or FALSE. Default value is FALSE.}

\item{urlproxy}{If Github package and TRUE proxy: proxy domain}

\item{portproxy}{If Github package and TRUE proxy: proxy port}

\item{force.install}{TRUE to force installation}
}
\description{
Download, install and load packages from Github repositories.
}
\examples{
## cran example: 
# inst.all(pckgName = "rpart")
# inst.all(pckgName = "rpart", force.install = TRUE)


# # bioconductor example: 
# inst.all(pckgName = "IRanges", bioc = TRUE)
# inst.all(pckgName = "IRanges", bioc = TRUE, force.install = TRUE)
 
 
# # github example: 
# inst.all(pckgName = "mmotaF", pathGit = "miriammota")
# inst.all(pckgName = "mmotaF", pathGit = "miriammota", force.install = TRUE)
}
\author{
Miriam Mota \email{miriam.mota@vhir.org}
}
\keyword{bioconductor}
\keyword{cran}
\keyword{github}
\keyword{install}
\keyword{library}
\keyword{require}
