% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installGitifnot.R
\name{installGitifnot}
\alias{installGitifnot}
\title{A installGitifnot Function}
\usage{
installGitifnot(pathGit, pckgName, proxy = FALSE,
  urlproxy = "conf_www.ir.vhebron.net", portproxy = 8081,
  force.install = FALSE)
}
\arguments{
\item{pathGit}{Repository address in the format username/repo[/subdir].}

\item{pckgName}{names of package}

\item{proxy}{TRUE or FALSE. Default value is FALSE.}

\item{urlproxy}{proxy domain}

\item{portproxy}{proxy port}

\item{force.install}{TRUE to force installation}
}
\description{
Download, install and load packages from Github repositories.
}
\examples{
## installGitifnot(pathGit = "miriammota",pckgName ="mmotaF",force.install = TRUE)
}
\author{
Miriam Mota \email{miriam.mota@vhir.org}
}
\keyword{github}
\keyword{install}
\keyword{library}
\keyword{require}
