% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installifnot.R
\name{installifnot}
\alias{installifnot}
\title{A installifnot Function}
\usage{
installifnot(pckgName, force.install = FALSE)
}
\arguments{
\item{pckgName}{names of package}

\item{force.install}{TRUE to force installation}
}
\description{
Download, install and load packages from CRAN-like repositories.
}
\examples{
## installifnot("rpart")
}
\author{
Miriam Mota \email{miriam.mota@vhir.org}
}
\keyword{install}
\keyword{library}
\keyword{require}
