% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installBiocifnot.R
\name{installBiocifnot}
\alias{installBiocifnot}
\title{A installBiocifnot Function}
\usage{
installBiocifnot(pckgName, force.install = FALSE)
}
\arguments{
\item{pckgName}{name of package}

\item{force.install}{TRUE to force installation}
}
\description{
Download, install and load packages from Bioconductor repositories.
}
\examples{
# installBiocifnot("CMA")
}
\author{
Miriam Mota \email{miriam.mota@vhir.org}
}
\keyword{bioconductor}
\keyword{install}
\keyword{library}
\keyword{require}
