% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installifnot.R
\name{installifnot}
\alias{installifnot}
\title{installifnot Function}
\usage{
installifnot(pkg)
}
\arguments{
\item{pkg}{Nombre del paquete en formato caracter}
}
\description{
Comprueba si un paquete esta instalado y si no lo instala y carga.
}
\examples{
installifnot("readxl")
}
\keyword{install}

